pyproject.toml
[tool.poetry]
name = "nico-making-airbnb-recap"
version = "0.1.0"
description = ""
authors = ["zelonKim <ksz18601@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1"
pillow = "^11.0.0"
djangorestframework = "^3.13.1"
strawberry-graphql = {version = "^0.130.4", extras = ["debug-server"]}
strawberry-graphql-django = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
